chore: don't add new button, use old one
